Add test. WIP. Requires merging PR#810
